base: Introduce Accidental Touch
Accidental Touch is a feature useful for many, especially the gamers out there.
It prevents any sort of accidental touch on hardware buttons while the touchscreen is
in use.
This means you won’t experience any unexpected behaviour while playing games,
web browsing or even when attempting to reach your finger out to the edge of
that humongous display.

Change-Id: I60c619ee04ec52c14f769adc76a7fe5a5782f906
Signed-off-by: Anush12133 <Anush.4376@gmail.com>
Signed-off-by: Srijith Bharadwaj <srijithdamerla2001@gmail.com>