Google FRP: save password quality in resetPassword
Since O MR1, when Google FRP is challenged, pattern/PIN/password which set
before erase can replace account verification. But in case of remote lock
by resetPassword, verification step is skipped.

Reproducible path:
1. Create Google account.
2. Remote lock & Erase in Google Find My Device.

The reason is because saving password quality is missed in resetPassword.
This patch is to fix that.

Change-Id: I54376f60ac53451ace22965d331b47cd8c2e614e
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Anush12133 <Anush.4376@gmail.com>
Signed-off-by: Srijith Bharadwaj <srijithdamerla2001@gmail.com>