base: Use a system-wide recents corner radius
With Google implementation of quickstep, the rounded corners are set indivdually according to the view.
If it's the AOSP default recents, it'll take a value for rounded corners.
If it's Launcher3 or any other app, it'll take a custom value that's declared on the app.

Let's make a public value and use it on both.

Change-Id: Ic91e3dd631f0466e1c6ff68755b1808fe698533d
Signed-off-by: Srijith Bharadwaj <srijithdamerla2001@gmail.com>